requirements.txt
ipython==7.5.0
lxml==4.3.4
python-telegram-bot==11.1.0
requests==2.22.0
cssselect==1.0.3
peewee==3.9.6
